Fixed Progress Bar
Current Progress bar progresses properly through shared folder "Data Backup" to find the file size, then measures how many files are completed as a measure of how close to complete the processing is. Currently does not reach 100%, as there are extra files in "Data Backup" that are not copied.

could rectify with an artificial 100%, or actually scan to not count duplicates (unlikely)